Remove unneeded transformer factory
